Minor simplifications and inlines

+++ src/main/java/org/jboss/ejb/client/remoting/ChannelAssociation.java
@@ -380,7 +380,7 @@ class ChannelAssociation {
         return messageCompatibilityChecker.isMessageCompatible(messageHeader);
     }
 
-    byte getNegotiatedProtocolVersion() {
+    int getNegotiatedProtocolVersion() {
         return this.protocolVersion;
     }
